+++ test/functional/transaction/transaction-decorator/controller/PostController.ts
@@ -1,13 +1,13 @@
 import {Transaction} from "../../../../../src/decorator/transaction/Transaction";
 import {Post} from "../entity/Post";
 import {EntityManager} from "../../../../../src/entity-manager/EntityManager";
-import {TransactionEntityManager} from "../../../../../src/decorator/transaction/TransactionEntityManager";
+import {TransactionManager} from "../../../../../src/decorator/transaction/TransactionManager";
 import {Category} from "../entity/Category";
 
 export class PostController {
 
     @Transaction("mysql") // "mysql" is a connection name. you can not pass it if you are using default connection.
-    async save(post: Post, category: Category, @TransactionEntityManager() entityManager: EntityManager) {
+    async save(post: Post, category: Category, @TransactionManager() entityManager: EntityManager) {
         await entityManager.save(post);
         await entityManager.save(category);
     }